:children_crossing: added request timeout for all calls
